Create Vector-Erase.cpp
The first line of the input contains an integer N. The next line contains N space-separated integers(1-based index). The third line contains a single integer x, denoting the position of an element that should be removed from the vector. The fourth line contains two integers a and b denoting the range that should be erased from the vector inclusive of a and exclusive of b.
Output Format
Print the size of the vector in the first line and the elements of the vector after the two erase operations in the second line separated by space.